pyproject.toml
[project]
name = "notebooklm_automator"
version = "0.2.0"
description = "FastAPI-based automation for Google NotebookLM using Playwright and CDP."
authors = [{ name = "Your Name", email = "your.email@example.com" }]
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.40.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "ulid-py>=1.1.0",
    "requests>=2.31.0"
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0"
]

[project.scripts]
run-server = "notebooklm_automator.main:main"
run-tests = "run_tests:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
notebooklm_automator = { path = "src/notebooklm_automator" }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-ra -q"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "e2e: marks tests as end-to-end (requires NOTEBOOKLM_URL)",
]

[tool.coverage.run]
source = ["src/notebooklm_automator"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
